Heed -Wconversion and -Wshadow warnings JIRA:CXX-5413
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@82541 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/connect/ncbi_misc.hpp
@@ -152,7 +152,9 @@ inline double CRateMonitor::GetTime(void) const
 
 inline double CRateMonitor::GetPace(void) const
 {
-    return GetTime() ? m_Data.front().first / m_Data.front().second : 0.0;
+    return GetTime()
+        ? double(m_Data.front().first) / m_Data.front().second
+        : 0.0;
 }
